release gate: make the half of it that is not test code pass
.wisent-release.json declares two quality gates for this product, `cargo fmt
--check` and `cargo clippy --all-targets -D warnings`, and neither had ever
passed. clippy was not installed for the pinned toolchain, so nobody had run
them locally; that is the practical reason stado has a release publisher, a
signing key and a manifest, and no releases.

Production code now passes both. Eleven clippy findings fixed rather than
silenced where silencing would hide something: three redundant `u32::from`
around `u8::BITS`, which the repository's own ban on bare numeric literals had
pushed people into writing; two `println!` headers with a trailing empty format
placeholder; one clone building a one-element slice. The five that are shape,
not defect -- three wide signatures and two enums with one large variant -- carry
an `#[allow]` with a sentence saying why the alternative is worse, because a
bare allow is the same unexplained declaration this week has been about.

The `cargo fmt --all` pass is why this commit touches 43 files. The tree had
never been formatted, so the gate could not pass without it, and running the
project's own formatter is the fix rather than a restyle.

What still fails is `--all-targets`, which compiles test code: `credential_store/tests.rs`
is missing `LazyLock` and `Mutex` imports, and three fixtures are missing fields
their structs gained -- `source_commit` on ReleaseRequest and ReleasePlan,
`identities` and `release_platform` on ComputeTarget, and `standby` and `verify`
on Service. Those are test-code repairs and are out of scope for this agent in
this workspace, so the gate stays red for a reason that is now named and small.

+++ stado-rs/src/cli/service.rs
@@ -1618,7 +1618,7 @@ async fn install_from_archive(
         hasher.update(&bytes);
         hex::encode(hasher.finalize())
     };
-    let version = format!("sha256-{}", &digest[..usize::from(u8::from(12u8))]);
+    let version = format!("sha256-{}", &digest[..usize::from(12u8)]);
     let staged = format!(".stado/.{directory}-{version}.tar.gz");
 
     if crate::deploy::host_channel::target_is_this_host(target) {
